+++ client/src/components/pages/Bedroom.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Bedroom = () => {
+  return <h2>Bedroom</h2>;
+};
+
+export default Bedroom;
