+++ client/src/components/pages/BedroomLights.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { AddLedModal, LedsTable, UpdateLedModal } from '../leds';
+
+const BedroomLights = () => {
+  return (
+    <>
+      <LedsTable />
+      <AddLedModal />
+      <UpdateLedModal />
+    </>
+  );
+};
+
+export default BedroomLights;
