+++ client/src/components/pages/Dashboard.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Dashboard = () => {
+  return <h2>Dashboard</h2>;
+};
+
+export default Dashboard;
