+++ client/src/components/pages/Home.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Home = () => {
+  return <h2>Home</h2>;
+};
+
+export default Home;
