+++ client/src/components/pages/NotFound.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const NotFound = () => {
+  return <h2>NotFound</h2>;
+};
+
+export default NotFound;
