+++ client/src/components/pages/index.js
@@ -1,0 +1,5 @@
+export { default as Bedroom } from './Bedroom';
+export { default as BedroomLights } from './BedroomLights';
+export { default as Dashboard } from './Dashboard';
+export { default as Home } from './Home';
+export { default as NotFound } from './NotFound';
